requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopy==2.1.0
idna==2.10
nominatim==0.1
numpy==1.20.0
pandas==1.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
ratelim==0.1.6
requests==2.25.1
six==1.15.0
soupsieve==2.1
urllib3==1.26.3
